requirements.txt
python-telegram-bot==20.*
google-generativeai
Pillow
